add sdp-jingle-json patch to fix calls to Conversations from Firefox
Firefox offers contain the "telephone-event" payload indicating DTMF support.
It also contains the correponding events parameter with no name, only "0-15".

Conversations does not allow payload parameters with no names and thus closes
the connection with "application-error". It makes sense since XEP-0167 says
that "Parameter names are effectively guaranteed to be unique".

Since names must be unique, an empty name should be invalid and dropped. The
best place seems to be in the sdp-jingle-json parsers.